move methods, convert to es6
