+++ src/utils/nimbledroid/index.jsx
@@ -1,5 +1,6 @@
 /* eslint-disable no-param-reassign */
 import CONFIG from './config';
+import { Exception } from '../../vendor/errors';
 
 export const sortSitesByTargetRatio = (a, b) => b.ratio - a.ratio;
 
@@ -8,7 +9,7 @@ const statusColor = (ratio, targetRatio) => {
   let widgetColor = 'green';
 
   if (targetRatio <= 1) {
-    throw new Error('Change the code to handle a ratio below 1.');
+    throw new Exception('Change the code to handle a ratio below 1.');
   }
 
   if (ratio > targetRatio) {
