+++ src/utils/perfherder/chartJs/getPerfherderData.js
@@ -1,8 +1,8 @@
 import { queryPerformanceData } from '../../../vendor/perf-goggles';
 import perfherderFormatter from './perfherderFormatter';
 import SETTINGS from '../../../settings';
-import { missing } from '../../../vendor/queryOps';
-import { Except } from '../../../vendor/utils/errors';
+import { missing } from '../../../vendor/utils';
+import { Exception } from '../../../vendor/errors';
 
 const getPerfherderData = async series => {
   const newData = new Array(series.length);
@@ -38,7 +38,7 @@ const getPerfherderData = async series => {
   );
 
   if (missing(newData[0]))
-    throw new Except(
+    throw new Exception(
       `can not data for ${JSON.stringify(series[0].seriesConfig)}`
     );
